+++ src/tag/parser/compile/scenario/post_compile.cpp
@@ -99,8 +99,8 @@ namespace Invader::Parser {
             // Figure out the base tag struct thing
             auto *bsp_tag_struct = &workload.structs[workload.tags[b.structure_bsp.tag_id.index].base_struct.value()];
             
-            // If we're not on dark circlet, we need to read the pointer at the beginning of the struct
-            if(workload.engine_target != HEK::CacheFileEngine::CACHE_FILE_DARK_CIRCLET) {
+            // If we're not on native, we need to read the pointer at the beginning of the struct
+            if(workload.engine_target != HEK::CacheFileEngine::CACHE_FILE_NATIVE) {
                 bsp_tag_struct = &workload.structs[bsp_tag_struct->resolve_pointer(static_cast<std::size_t>(0)).value()];
             }
 
@@ -554,8 +554,8 @@ namespace Invader::Parser {
                     // Figure out the base tag struct thing
                     auto *bsp_tag_struct = &workload.structs[workload.tags[bsp_id.index].base_struct.value()];
                     
-                    // If we're not on dark circlet, we need to read the pointer at the beginning of the struct
-                    if(workload.engine_target != HEK::CacheFileEngine::CACHE_FILE_DARK_CIRCLET) {
+                    // If we're not on native, we need to read the pointer at the beginning of the struct
+                    if(workload.engine_target != HEK::CacheFileEngine::CACHE_FILE_NATIVE) {
                         bsp_tag_struct = &workload.structs[bsp_tag_struct->resolve_pointer(static_cast<std::size_t>(0)).value()];
                     }